+++ CHANGELOG.md
@@ -14,6 +14,12 @@ and this project adheres to
 
 ### Changed
 
+- Updated @openfn/workflow-diagram to 0.4.0
+- Make plus button part of job nodes in Workflow Diagram
+- Updated @openfn/adaptor-docs to 0.0.5
+- Updated @openfn/describe-package to 0.0.10
+- Create an follow a manual Run from the Job Inspector
+
 ### Fixed
 
 ## [0.3.1] - 2022-11-22